+++ check/file.go
@@ -1,6 +1,9 @@
 package check
 
 import (
+	"fmt"
+	"log"
+	"os"
 	"path/filepath"
 )
 
@@ -19,4 +22,20 @@ func (opts *FileOptions) FullPath(path string) string {
 	}
 
 	return path
+}
+
+// FileSizeCheck verifies that documentation file is below the Terraform Registry storage limit.
+func FileSizeCheck(fullpath string) error {
+	fi, err := os.Stat(fullpath)
+
+	if err != nil {
+		return err
+	}
+
+	log.Printf("[DEBUG] File %s size: %d (limit: %d)", fullpath, fi.Size(), RegistryMaximumSizeOfFile)
+	if fi.Size() >= int64(RegistryMaximumSizeOfFile) {
+		return fmt.Errorf("exceeded maximum (%d) size of documentation file for Terraform Registry: %d", RegistryMaximumSizeOfFile, fi.Size())
+	}
+
+	return nil
 }