+++ check/registry_data_source_file.go
@@ -48,7 +48,11 @@ func (check *RegistryDataSourceFileCheck) Run(path string) error {
 	log.Printf("[DEBUG] Checking file: %s", fullpath)
 
 	if err := RegistryFileExtensionCheck(path); err != nil {
-		return fmt.Errorf("error checking file (%s) extension: %w", path, err)
+		return fmt.Errorf("%s: error checking file extension: %w", path, err)
+	}
+
+	if err := FileSizeCheck(fullpath); err != nil {
+		return fmt.Errorf("%s: error checking file size: %w", path, err)
 	}
 
 	content, err := ioutil.ReadFile(fullpath)